+++ sdk/storage/storage-queue/src/QueueClient.ts
@@ -72,6 +72,23 @@ export interface QueueCreateOptions extends CommonOptions {
 }
 
 /**
+ * Options to configure {@link QueueClient.exists} operation
+ *
+ * @export
+ * @interface QueueExistsOptions
+ */
+export interface QueueExistsOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   *
+   * @type {AbortSignalLike}
+   * @memberof QueueExistsOptions
+   */
+  abortSignal?: AbortSignalLike;
+}
+
+/**
  * Options to configure {@link QueueClient.getProperties} operation
  *
  * @export
@@ -193,24 +210,24 @@ export interface SignedIdentifier {
 export declare type QueueGetAccessPolicyResponse = {
   signedIdentifiers: SignedIdentifier[];
 } & QueueGetAccessPolicyHeaders & {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpResponse & {
     /**
-     * The underlying HTTP response.
+     * The parsed HTTP response headers.
      */
-    _response: HttpResponse & {
-      /**
-       * The parsed HTTP response headers.
-       */
-      parsedHeaders: QueueGetAccessPolicyHeaders;
-      /**
-       * The response body as text (string format)
-       */
-      bodyAsText: string;
-      /**
-       * The response body as parsed JSON or XML
-       */
-      parsedBody: SignedIdentifierModel[];
-    };
+    parsedHeaders: QueueGetAccessPolicyHeaders;
+    /**
+     * The response body as text (string format)
+     */
+    bodyAsText: string;
+    /**
+     * The response body as parsed JSON or XML
+     */
+    parsedBody: SignedIdentifierModel[];
   };
+};
 
 /**
  * Options to configure {@link QueueClient.clearMessages} operation
@@ -236,7 +253,7 @@ export interface QueueClearMessagesOptions extends CommonOptions {
  * @interface QueueSendMessageOptions
  * @extends {MessagesEnqueueOptionalParams}
  */
-export interface QueueSendMessageOptions extends MessagesEnqueueOptionalParams, CommonOptions {}
+export interface QueueSendMessageOptions extends MessagesEnqueueOptionalParams, CommonOptions { }
 
 /**
  * Options to configure {@link QueueClient.receiveMessages} operation
@@ -245,7 +262,7 @@ export interface QueueSendMessageOptions extends MessagesEnqueueOptionalParams, 
  * @interface QueueReceiveMessageOptions
  * @extends {MessagesDequeueOptionalParams}
  */
-export interface QueueReceiveMessageOptions extends MessagesDequeueOptionalParams, CommonOptions {}
+export interface QueueReceiveMessageOptions extends MessagesDequeueOptionalParams, CommonOptions { }
 
 /**
  * Options to configure {@link QueueClient.peekMessages} operation
@@ -254,7 +271,7 @@ export interface QueueReceiveMessageOptions extends MessagesDequeueOptionalParam
  * @interface QueuePeekMessagesOptions
  * @extends {MessagesPeekOptionalParams}
  */
-export interface QueuePeekMessagesOptions extends MessagesPeekOptionalParams, CommonOptions {}
+export interface QueuePeekMessagesOptions extends MessagesPeekOptionalParams, CommonOptions { }
 
 /**
  * Contains the response data for the {@link QueueClient.sendMessage} operation.
@@ -286,24 +303,24 @@ export declare type QueueSendMessageResponse = {
    */
   nextVisibleOn: Date;
 } & MessagesEnqueueHeaders & {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpResponse & {
     /**
-     * The underlying HTTP response.
+     * The parsed HTTP response headers.
      */
-    _response: HttpResponse & {
-      /**
-       * The parsed HTTP response headers.
-       */
-      parsedHeaders: MessagesEnqueueHeaders;
-      /**
-       * The response body as text (string format)
-       */
-      bodyAsText: string;
-      /**
-       * The response body as parsed JSON or XML
-       */
-      parsedBody: EnqueuedMessage[];
-    };
+    parsedHeaders: MessagesEnqueueHeaders;
+    /**
+     * The response body as text (string format)
+     */
+    bodyAsText: string;
+    /**
+     * The response body as parsed JSON or XML
+     */
+    parsedBody: EnqueuedMessage[];
   };
+};
 
 /**
  * The object returned in the `receivedMessageItems` array when calling {@link QueueClient.receiveMessages}.
@@ -318,24 +335,24 @@ export declare type ReceivedMessageItem = DequeuedMessageItem;
 export declare type QueueReceiveMessageResponse = {
   receivedMessageItems: ReceivedMessageItem[];
 } & MessagesDequeueHeaders & {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpResponse & {
     /**
-     * The underlying HTTP response.
+     * The parsed HTTP response headers.
      */
-    _response: HttpResponse & {
-      /**
-       * The parsed HTTP response headers.
-       */
-      parsedHeaders: MessagesDequeueHeaders;
-      /**
-       * The response body as text (string format)
-       */
-      bodyAsText: string;
-      /**
-       * The response body as parsed JSON or XML
-       */
-      parsedBody: ReceivedMessageItem[];
-    };
+    parsedHeaders: MessagesDequeueHeaders;
+    /**
+     * The response body as text (string format)
+     */
+    bodyAsText: string;
+    /**
+     * The response body as parsed JSON or XML
+     */
+    parsedBody: ReceivedMessageItem[];
   };
+};
 
 /**
  * Contains the response data for the {@link QueueClient.peekMessages} operation.
@@ -343,24 +360,24 @@ export declare type QueueReceiveMessageResponse = {
 export declare type QueuePeekMessagesResponse = {
   peekedMessageItems: PeekedMessageItem[];
 } & MessagesPeekHeaders & {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpResponse & {
     /**
-     * The underlying HTTP response.
+     * The parsed HTTP response headers.
      */
-    _response: HttpResponse & {
-      /**
-       * The parsed HTTP response headers.
-       */
-      parsedHeaders: MessagesPeekHeaders;
-      /**
-       * The response body as text (string format)
-       */
-      bodyAsText: string;
-      /**
-       * The response body as parsed JSON or XML
-       */
-      parsedBody: PeekedMessageItem[];
-    };
+    parsedHeaders: MessagesPeekHeaders;
+    /**
+     * The response body as text (string format)
+     */
+    bodyAsText: string;
+    /**
+     * The response body as parsed JSON or XML
+     */
+    parsedBody: PeekedMessageItem[];
   };
+};
 
 /**
  * Options to configure the {@link QueueClient.deleteMessage} operation
@@ -609,6 +626,81 @@ export class QueueClient extends StorageClient {
   }
 
   /**
+   * Creates a new queue under the specified account if it doesn't already exist.
+   * If the queue already exists, it is not changed.
+   * @see https://docs.microsoft.com/en-us/rest/api/storageservices/create-queue4
+   *
+   * @param {QueueCreateOptions} [options]
+   * @returns {Promise<QueueCreateResponse | null>} Returns null if the queue already exists.
+   * @memberof QueueClient
+   */
+  public async createIfNotExists(options: QueueCreateOptions = {}): Promise<QueueCreateResponse | null> {
+    const { span, spanOptions } = createSpan("QueueClient-createIfNotExists", options.tracingOptions);
+    try {
+      const response = await this.create({
+        ...options,
+        tracingOptions: { ...options!.tracingOptions, spanOptions }
+      });
+
+      // When a queue with the specified name already exists, the Queue service checks the metadata associated with the existing queue.
+      // If the existing metadata is identical to the metadata specified on the Create Queue request, status code 204 (No Content) is returned.
+      // If the existing metadata does not match, the operation fails and status code 409 (Conflict) is returned.
+      if (response._response.status == 204) {
+        return null;
+      }
+      return response;
+    } catch (e) {
+      if (e.details?.errorCode === "QueueAlreadyExists") {
+        span.setStatus({
+          code: CanonicalCode.ALREADY_EXISTS,
+          message: "Expected exception when creating a queue only if it does not already exist."
+        });
+        return null;
+      }
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Deletes the specified queue permanently if it exists.
+   * @see https://docs.microsoft.com/en-us/rest/api/storageservices/delete-queue3
+   *
+   * @param {QueueDeleteOptions} [options]
+   * @returns {Promise<QueueDeleteResponse | null>} Returns null if the queue doesn't exists.
+   * @memberof QueueClient
+   */
+  public async deleteIfExists(options: QueueDeleteOptions = {}): Promise<QueueDeleteResponse | null> {
+    const { span, spanOptions } = createSpan("QueueClient-deleteIfExists", options.tracingOptions);
+    try {
+      return await this.delete({
+        ...options,
+        tracingOptions: { ...options!.tracingOptions, spanOptions }
+      });
+    } catch (e) {
+      if (e.details?.errorCode === "QueueNotFound") {
+        span.setStatus({
+          code: CanonicalCode.NOT_FOUND,
+          message: "Expected exception when deleting a queue only if it exists."
+        });
+        return null;
+      }
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
    * Deletes the specified queue permanently.
    * @see https://docs.microsoft.com/en-us/rest/api/storageservices/delete-queue3
    *
@@ -633,6 +725,43 @@ export class QueueClient extends StorageClient {
         spanOptions
       });
     } catch (e) {
+      span.setStatus({
+        code: CanonicalCode.UNKNOWN,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Returns true if the specified queue exists; false otherwise.
+   *
+   * NOTE: use this function with care since an existing queue might be deleted by other clients or
+   * applications. Vice versa new queues might be added by other clients or applications after this
+   * function completes.
+   *
+   * @param {QueueExistsOptions} [options] options to Exists operation.
+   * @returns {Promise<boolean>}
+   * @memberof QueueClient
+   */
+  public async exists(options: QueueExistsOptions = {}): Promise<boolean> {
+    const { span, spanOptions } = createSpan("QueueClient-exists", options.tracingOptions);
+    try {
+      await this.getProperties({
+        abortSignal: options.abortSignal,
+        tracingOptions: { ...options.tracingOptions, spanOptions }
+      });
+      return true;
+    } catch (e) {
+      if (e.statusCode === 404) {
+        span.setStatus({
+          code: CanonicalCode.NOT_FOUND,
+          message: "Expected exception when checking queue existence"
+        });
+        return false;
+      }
       span.setStatus({
         code: CanonicalCode.UNKNOWN,
         message: e.message