+++ sdk/storage/storage-blob/test/appendblobclient.spec.ts
@@ -64,6 +64,16 @@ describe("AppendBlobClient", () => {
     assert.equal(properties.metadata!.key2, options.metadata.key2);
   });
 
+  it("createIfNotExists", async () => {
+    const res = await appendBlobClient.createIfNotExists();
+    assert.ok(res.succeeded);
+    assert.ok(res.etag);
+
+    const res2 = await appendBlobClient.createIfNotExists();
+    assert.ok(!res2.succeeded);
+    assert.equal(res2.errorCode, "BlobAlreadyExists");
+  });
+
   it("appendBlock", async () => {
     await appendBlobClient.create();